Move validators into separate modules
