+++ src/dsl/validation.rs
@@ -1,211 +1,12 @@
-use super::types::ValidationError;
+use super::types::{ValidationError, Validator};
+use crate::validators::expectation_validator::ExpectationValidator;
+use crate::validators::schema_validator::SchemaValidator;
+use crate::validators::value_validator::ValueValidator;
 use colored::*;
 use jsonschema::{Draft, JSONSchema};
-use rhai::{Engine, Expr, Stmt};
-use serde_json::json;
+use rhai::Engine;
 
 const SCHEMA: &str = include_str!("../../wanda/guides/check_definition.schema.json");
-
-trait Validator {
-    fn validate(&self, json_check: &serde_json::Value, check_id: &str) -> Vec<ValidationError>;
-}
-
-struct SchemaValidator<'a> {
-    schema: &'a JSONSchema,
-}
-
-impl<'a> Validator for SchemaValidator<'a> {
-    fn validate(&self, json_check: &serde_json::Value, check_id: &str) -> Vec<ValidationError> {
-        validate_schema(json_check, check_id, self.schema)
-    }
-}
-
-struct ExpectationValidator<'a> {
-    engine: &'a Engine,
-}
-
-impl<'a> Validator for ExpectationValidator<'a> {
-    fn validate(&self, json_check: &serde_json::Value, check_id: &str) -> Vec<ValidationError> {
-        validate_expectations(json_check, check_id, self.engine)
-    }
-}
-
-struct ValueValidator<'a> {
-    engine: &'a Engine,
-}
-
-impl<'a> Validator for ValueValidator<'a> {
-    fn validate(&self, json_check: &serde_json::Value, check_id: &str) -> Vec<ValidationError> {
-        validate_values(json_check, check_id, self.engine)
-    }
-}
-
-fn validate_schema(
-    json_check: &serde_json::Value,
-    check_id: &str,
-    schema: &JSONSchema,
-) -> Vec<ValidationError> {
-    let validation_result = match schema.validate(json_check) {
-        Ok(_) => Ok(()),
-        Err(errors) => {
-            let validation_errors = errors
-                .map(|error| ValidationError {
-                    check_id: check_id.to_string(),
-                    error: error.to_string(),
-                    instance_path: error.instance_path.to_string(),
-                })
-                .collect();
-            Err(validation_errors)
-        }
-    };
-
-    return match validation_result {
-        Ok(_) => vec![],
-        Err(errors) => errors,
-    };
-}
-
-fn validate_expectations(
-    json_check: &serde_json::Value,
-    check_id: &str,
-    engine: &Engine,
-) -> Vec<ValidationError> {
-    let (_, expectation_expression_errors): (Vec<_>, Vec<_>) = json_check
-        .get("expectations")
-        .unwrap_or(&json!([]))
-        .as_array()
-        .unwrap_or(&Vec::new())
-        .iter()
-        .enumerate()
-        .flat_map(|(index, value)| {
-            let expect = value.get("expect");
-            let expect_same = value.get("expect_same");
-            let expect_enum = value.get("expect_enum");
-
-            let is_expect = expect.is_some();
-            let is_expect_same = expect_same.is_some();
-            let is_expect_enum = expect_enum.is_some();
-
-            let expectation_expression = if is_expect {
-                expect.unwrap().as_str().unwrap()
-            } else if is_expect_same {
-                expect_same.unwrap().as_str().unwrap()
-            } else if is_expect_enum {
-                expect_enum.unwrap().as_str().unwrap()
-            } else {
-                ""
-            };
-
-            let mut results = vec![];
-
-            match engine.compile(expectation_expression) {
-                Ok(_) => results.push(Ok(())),
-                Err(error) => results.push(Err(ValidationError {
-                    check_id: check_id.to_string(),
-                    error: error.to_string(),
-                    instance_path: format!("/expectations/{:?}", index).to_string(),
-                })),
-            }
-
-            let failure_message = value.get("failure_message");
-            let warning_message = value.get("warning_message");
-
-            if failure_message.is_some() {
-                let failure_message_expression = failure_message.unwrap().as_str().unwrap();
-                results.push(validate_string_expression(
-                    failure_message_expression,
-                    engine,
-                    check_id,
-                    index,
-                    is_expect || is_expect_enum,
-                ));
-            }
-
-            if warning_message.is_some() && !is_expect_enum {
-                results.push(Err(ValidationError {
-                    check_id: check_id.to_string(),
-                    error: "warning_message is only available for expect_enum expectations"
-                        .to_string(),
-                    instance_path: format!("/expectations/{:?}", index).to_string(),
-                }));
-            } else if warning_message.is_some() {
-                let warning_message_expression = warning_message.unwrap().as_str().unwrap();
-                results.push(validate_string_expression(
-                    warning_message_expression,
-                    engine,
-                    check_id,
-                    index,
-                    is_expect_enum,
-                ));
-            }
-
-            if is_expect_enum {
-                results.append(&mut validate_expect_enum_content(
-                    expectation_expression,
-                    check_id,
-                    index,
-                ));
-            }
-
-            results
-        })
-        .partition(Result::is_ok);
-
-    return expectation_expression_errors
-        .into_iter()
-        .map(Result::unwrap_err)
-        .collect();
-}
-
-fn validate_values(
-    json_check: &serde_json::Value,
-    check_id: &str,
-    engine: &Engine,
-) -> Vec<ValidationError> {
-    let (_, values_expression_errors): (Vec<_>, Vec<_>) = json_check
-        .get("values")
-        .unwrap_or(&json!([]))
-        .as_array()
-        .unwrap_or(&Vec::new())
-        .iter()
-        .enumerate()
-        .flat_map(|(value_index, value)| {
-            let conditions_compilations_results: Vec<Result<_, _>> = value
-                .get("conditions")
-                .unwrap_or(&json!([]))
-                .as_array()
-                .unwrap_or(&Vec::new())
-                .iter()
-                .enumerate()
-                .map(|(condition_index, condition)| {
-                    let default_json_string = json!("");
-                    let when_expression = condition
-                        .get("when")
-                        .unwrap_or(&default_json_string)
-                        .as_str()
-                        .unwrap();
-                    engine
-                        .compile(when_expression)
-                        .map_err(|error| ValidationError {
-                            check_id: check_id.to_string(),
-                            error: error.to_string(),
-                            instance_path: format!(
-                                "/values/{:?}/conditions/{:?}",
-                                value_index, condition_index
-                            ),
-                        })
-                })
-                .collect();
-
-            conditions_compilations_results
-        })
-        .partition(Result::is_ok);
-
-    return values_expression_errors
-        .into_iter()
-        .map(Result::unwrap_err)
-        .collect();
-}
 
 pub fn error_header(head: &str) -> String {
     format!("  {}  ", head).on_red().black().to_string()
@@ -234,85 +35,6 @@ pub fn validate(
     }
 
     Err(errors)
-}
-
-fn validate_string_expression(
-    expression: &str,
-    engine: &Engine,
-    check_id: &str,
-    index: usize,
-    allow_interpolated_strings: bool,
-) -> Result<(), ValidationError> {
-    match engine.compile(format!("`{}`", expression)) {
-        Ok(ast) => {
-            let statements = ast.statements();
-            if statements.len() > 1 {
-                return Err(ValidationError {
-                    check_id: check_id.to_string(),
-                    error: "Too many statements".to_string(),
-                    instance_path: format!("/expectations/{:?}", index).to_string(),
-                });
-            }
-
-            match &statements[0] {
-                Stmt::Expr(expression) => match **expression {
-                    Expr::StringConstant(_, _) => Ok(()),
-                    Expr::InterpolatedString(_, _) => {
-                        if !allow_interpolated_strings {
-                            Err(ValidationError {
-                                check_id: check_id.to_string(),
-                                error: "String interpolation is not allowed here".to_string(),
-                                instance_path: format!("/expectations/{:?}", index).to_string(),
-                            })
-                        } else {
-                            Ok(())
-                        }
-                    }
-                    _ => Err(ValidationError {
-                        check_id: check_id.to_string(),
-                        error: "Field has to be a string".to_string(),
-                        instance_path: format!("/expectations/{:?}", index).to_string(),
-                    }),
-                },
-                _ => Err(ValidationError {
-                    check_id: check_id.to_string(),
-                    error: "Field has to be an expression".to_string(),
-                    instance_path: format!("/expectations/{:?}", index).to_string(),
-                }),
-            }
-        }
-        Err(error) => Err(ValidationError {
-            check_id: check_id.to_string(),
-            error: error.to_string(),
-            instance_path: format!("/expectations/{:?}", index).to_string(),
-        }),
-    }
-}
-
-fn validate_expect_enum_content(
-    expression: &str,
-    check_id: &str,
-    index: usize,
-) -> Vec<Result<(), ValidationError>> {
-    let mut results = vec![];
-
-    if !expression.contains("\"passing\"") {
-        results.push(Err(ValidationError {
-            check_id: check_id.to_string(),
-            error: "passing return value not found".to_string(),
-            instance_path: format!("/expectations/{:?}", index).to_string(),
-        }));
-    }
-
-    if !expression.contains("\"warning\"") {
-        results.push(Err(ValidationError {
-      check_id: check_id.to_string(),
-      error: "warning return value not found. Consider using `expect` expression if a warning return is not needed".to_string(),
-      instance_path: format!("/expectations/{:?}", index).to_string(),
-    }));
-    }
-
-    results
 }
 
 pub fn get_json_schema() -> JSONSchema {
