+++ src/validators/schema_validator.rs
@@ -1,0 +1,134 @@
+use crate::dsl::types::{ValidationError, Validator};
+use jsonschema::JSONSchema;
+
+pub struct SchemaValidator<'a> {
+    pub schema: &'a JSONSchema,
+}
+
+impl<'a> Validator for SchemaValidator<'a> {
+    fn validate(&self, json_check: &serde_json::Value, check_id: &str) -> Vec<ValidationError> {
+        validate_schema(json_check, check_id, self.schema)
+    }
+}
+
+fn validate_schema(
+    json_check: &serde_json::Value,
+    check_id: &str,
+    schema: &JSONSchema,
+) -> Vec<ValidationError> {
+    let validation_result = match schema.validate(json_check) {
+        Ok(_) => Ok(()),
+        Err(errors) => {
+            let validation_errors = errors
+                .map(|error| ValidationError {
+                    check_id: check_id.to_string(),
+                    error: error.to_string(),
+                    instance_path: error.instance_path.to_string(),
+                })
+                .collect();
+            Err(validation_errors)
+        }
+    };
+
+    return match validation_result {
+        Ok(_) => vec![],
+        Err(errors) => errors,
+    };
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::dsl::types::Check;
+    use crate::dsl::validation::get_json_schema;
+    use serde_json;
+
+    #[test]
+    fn validate_wrong_check() {
+        let input = r#"
+            id: 156F64
+            name: Corosync configuration file
+            group: Corosync
+            description: |
+              Corosync `token` timeout is set to expected value
+            remediation: |
+              ## Abstract
+              The value of the Corosync `token` timeout is not set as recommended.
+              ## Remediation
+              ...
+            facts:
+              - name: corosync_token_timeout
+                gatherer: corosync.conf
+                argument: totem.token
+            values:
+              - name: expected_token_timeout
+                default: 5000
+                conditions:
+                  - value: 30000
+                    when: env.provider == "azure" || env.provider == "aws"
+                  - value: 20000
+                    whens: env.provider == "gcp"
+            expectations:
+              - name: timeout
+                expect: facts.corosync_token_timeout == values.expected_token_timeout
+        "#;
+
+        let json_value: serde_json::Value =
+            serde_yaml::from_str(input).expect("Unable to parse yaml");
+        let json_schema = get_json_schema();
+        let validation_errors = validate_schema(&json_value, "156F64", &json_schema);
+        assert_eq!(validation_errors[0].check_id, "156F64");
+        assert_eq!(
+            validation_errors[0].error,
+            "Additional properties are not allowed ('whens' was unexpected)"
+        );
+        assert_eq!(validation_errors[0].instance_path, "/values/0/conditions/1");
+    }
+
+    #[test]
+    fn validate_ok_check() {
+        let input = r#"
+            id: 156F64
+            name: Corosync configuration file
+            group: Corosync
+            description: |
+              Corosync `token` timeout is set to expected value
+            remediation: |
+              ## Abstract
+              The value of the Corosync `token` timeout is not set as recommended.
+              ## Remediation
+              ...
+            premium: true
+            metadata:
+              target_type: cluster
+              provider:
+                - aws
+                - azure
+            facts:
+              - name: corosync_token_timeout
+                gatherer: corosync.conf
+                argument: totem.token
+            values:
+              - name: expected_token_timeout
+                default: 5000
+                conditions:
+                  - value: 30000
+                    when: env.provider == "azure" || env.provider == "aws"
+                  - value: 20000
+                    when: env.provider == "gcp"
+            expectations:
+              - name: timeout
+                expect: facts.corosync_token_timeout == values.expected_token_timeout
+        "#;
+
+        let json_value: serde_json::Value =
+            serde_yaml::from_str(input).expect("Unable to parse yaml");
+        let json_schema = get_json_schema();
+        let validation_result = validate_schema(&json_value, "156F64", &json_schema);
+
+        let deserialization_result = serde_yaml::from_str::<Check>(input);
+
+        assert!(validation_result.is_empty());
+        assert!(deserialization_result.is_ok());
+    }
+}
